+++ CHANGELOG.md
@@ -57,6 +57,9 @@ In order to build the DUNE core modules you need at least the following software
 - Add the utility `IsCompileTimeConstant` to check for integral constants and anything with
   the same interface.
 
+- `DynamicMatrix` is now fully compatible to `FieldMatrix`. Two
+  methods (`rightmultiplyany` and `leftmultiplyany`) were added.
+
 ## Build System
 
 - Documentation files in `doc/buildsystem/${ModuleName}.rst` are now only copied.